Define kernel-vanilla as source variant
The vanilla_only macro is overloaded. It is used for determining if
there should be two kernel sources built as well as for the purpose of
determmioning if vanilla kernel should be used for kernel-obs-build.

While the former can be determined at build time the latter needs to be
baked into the spec file template. Separate the two while also making
the latter more generic.

$build_dtbs is enabled on every single rt and azure branch since 15.3
when the setting was introduced, gate on the new $obs_build_variant
setting as well.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>