scripts/bugzilla: Add heuristics for version selection
Product versions are not sorted chronologically (fun fact: in SLE12-SP5
lexicographical sort equaled chronological). The script workload doesn't
care about exact version, so use heuristics of a '*Maint-Upd'
maintenance update and fall back to 'unspecified' if available.
The goal is to supply a version that allows opening a new bug.
When the script needs to be used with finer version granularity, it must
be modified.

Signed-off-by: Michal Koutný <mkoutny@suse.com>