scripts/gitlog2changes: Handle SSH signatures
Commit a384f306f91 (Fix parsing of GPG-signed commit) added the ability
to handle lines beginning with gpgsig but only added the check for the
PGP signatures.  It would mark the state as being within a signature and
not print anything and get stuck in that state because the check was
only looking for PGP and not SSH signatures like the ones used in the
repo.