scripts/CKC: mark local variables as such
The default global and dynamic scope nature of bash variables is causing
some race conditions.  For example, missing hashes are sometimes printed
and sometimes not, depending on what is found in $missing variable.  For
loops and functions are polluting global namespace with outdated state
that is being picked up on their next run.  We should religiously mark
local variables as such unless we want to explicity do global store.

Signed-off-by: Miroslav Franc <mfranc@suse.cz>