Update config files: make SCSI/ATA drivers modules again
As discussed on opensuse-kernel ML, we want to make SCSI and ATA
drivers from built-in back to modules again:
  https://lists.opensuse.org/archives/list/kernel@lists.opensuse.org/thread/MLRQW7RFEAKTAP63NMPFFIYTXAF7E3I3/
They were made as built-in many many years ago just for boot speed up
and a slight hope of initrd-less systems.  But it makes more sense to
align with the SLE configurations.