Fix patches.suse/io_uring-used-cached-copies-of-sq-dropped-and-cq-ove.patch. (bsc#1214344)
To protect itself against userspace corrupting the counter of io_uring
dropped submission entries, the kernel relies on a cache of the counter
instead of reading the counter directly.  But, the stable patch that was
brought to SP3 implementing the this mechanism was done incorrectly, and
let's the kernel read from the userspace value instead of the cache in
one situation. This allows userspace to subvert the counter, hanging the
application forever. Fix the backport to read from the cached value.

5.3 stable is long dead, so there is nothing to fix upstream or in
-stable.