config: enable NFS_V4_2_READ_PLUS (bsc#1216736)
This option was introduced and disabled by default because of unresolved
issues. As stated in mainline commit 9cf2744d2491 ("NFS: Enable the
READ_PLUS operation by default") which changes the upstream default to
enabled, these issues are addressed so that it is safe to enable the
option and as it allows to transfer sparse files more efficiently, it is
desirable to do so.