rpm/kernel-source.changes.old: Remove old changelog entries
The RPM changelog has grown to ~400k lines which unnecessarily increases
size of the resulting packages, especially because the same changelog is
at the moment used for all built packages. For instance, size of the
kernel-macros package is now at ~12MB while the actual content is ~30kB.

To reduce the size, bump the cut-off commit for the changelog to the
predecessor of 126b9a7394d ("Change to SLE15-SP6 branch
(jsc#PED-4593)").

Note that the auto-generated changelog currently relies on author dates.
Some commits that were cherry-picked on the branch have their author
date earlier than the commit 126b9a7394d and hence appear before this
commit in the produced changelog as well.