Do not store build host name in initrd
Without this patch, kernel-obs-build stored the build host name
in its .build.initrd.kvm

This patch allows for reproducible builds of kernel-obs-build and thus
avoids re-publishing the kernel-obs-build.rpm when nothing changed.

Note that this has no influence on the /etc/hosts file
that is used during other OBS builds.

https://bugzilla.opensuse.org/show_bug.cgi?id=1084909

Signed-off-by: Bernhard M. Wiedemann <bwiedemann@suse.de>