scripts: pre-commit: Check newly added blacklist.conf entries
When blacklist.conf entries are added by merging an "upstream" branch,
they are not checked against present commits and the repo ends up in
inconsistent state when the patch is present and blacklisted at the same
time.

The state is checked in pre-commit hook when a (blacklisted) patch is
added. Prevent reaching this state when adding blacklist.conf entries
too.

Using scripts/check-patch-blacklist for this check would be
prohibitively slow (~5 minutes with 40k patches).