scripts: Add commit-msg check for patch references
References in the commit message are important when generating the RPM
changelog.
Although scripts/log takes into account References: header, a reference
may possibly be missed out when the script is skipped or the message
misedited.

Add a new hook that validates that the commit message contains all newly
added references.

Signed-off-by: Michal Koutný <mkoutny@suse.com>