rpm/constraints.in: add static multibuild packages
Commit 841012b049a5 (rpm/mkspec: use kernel-source: prefix for
constraints on multibuild) added "kernel-source:" prefix to the
dynamically generated kernels. But there are also static ones like
kernel-docs. Those fail to build as the constraints are still not
applied.

So add the prefix also to the static ones.

Note kernel-docs-rt is given kernel-source-rt prefix. I am not sure it
will ever be multibuilt...