rpm templates: Move macro definitions below buildrequires
Many of the rpm macros defined in the kernel packages depend directly or
indirectly on script execution. OBS cannot execute scripts which means
values of these macros cannot be used in tags that are required for OBS
to see such as package name, buildrequires or buildarch.

Accumulate macro definitions that are not directly expanded by mkspec
below buildrequires and buildarch to make this distinction clear.