scripts/check-kernel-fix: Enhancements and cleanups
Allow to check which branches have a given reference without
passing a particular sha. Actions are not printed in this case.

Show actions when "sha" is passed by default.

Add [-q] option to do not show progress when checking
state of each branch.

Do not show action for a branch when the merge branch
already has the patch with all references.

Check each branch only once for the given sha and all references.
It allowed to reduce git grep calls. Also it removed the need
to merge branch states. It improved speed and simplified
the logic.

Signed-off-by: Petr Mladek <pmladek@suse.com>