rpm/kernel-obs-build.spec.in: remove reiserfs from OBS initrd
We disabled the FS in bug 1202309. And we actively blacklist it in:
/usr/lib/modprobe.d/60-blacklist_fs-reiserfs.conf

This, as a side-effect, fixes obs-build's warning:
dracut-pre-udev[1463]: sh: line 1: /usr/lib/module-init-tools/unblacklist: No such file or directory

Exactly due to the above 60-blacklist_fs-reiserfs.conf trying to call the
above unblacklist.

We should likely drop ext2+ext3 from the list too, as we don't build
them at all. But that's a different story.