scripts/check-kernel-fix: print link to upstream commit in the summary
Initial CVE reports contain information assembled from various sources
and lack the upstream commit that's supposed to be backported. This
will not be fixed (https://gitlab.suse.de/tools/smash/-/issues/1601).

The bugs are usually updated with another comment with check-kernel-fix
so let's put the link there and it will be clickable when viewed on
bugzilla page. The output will look like:

  ./scripts/check-kernel-fix -s 0 CVE-2024-35949
  e03418abde87 ("btrfs: make sure that WRITTEN is set on all metadata blocks") merged v6.9~23^2
  No Fixes tag. Requires manual review for affected branches.
  Security fix for CVE-2024-35949 bsc#1224700 with CVSS 0
  Experts candidates: dsterba@suse.com
  Link: https://git.kernel.org/linus/e03418abde871314e1a3a550f4c8afb7b89cb273
  .............................
  ACTION NEEDED!
  SLE12-SP5: MANUAL: might need backport of e03418abde871314e1a3a550f4c8afb7b89cb273 ()
  SLE15-SP6: MANUAL: might need backport of e03418abde871314e1a3a550f4c8afb7b89cb273 ()
  SLE12-SP3-TD: MANUAL: might need backport of e03418abde871314e1a3a550f4c8afb7b89cb273 ()
  SLE15-SP5: MANUAL: might need backport of e03418abde871314e1a3a550f4c8afb7b89cb273 ()

i.e. line with "Link:".

Signed-off-by: David Sterba <dsterba@suse.com>