packaging: Add case-sensitive perl option parsing
A recent change in Getopt::Long [1]:

Changes in version 2.55
-----------------------

* Fix long standing bug that duplicate options were not detected
when the options differ in case while ignore_case is in effect.
This will now yield a warning and become a fatal error in a future
release.

perl defaults to ignore_case by default, switch it off to avoid
accidental misparsing of options.
This was suggested after similar change in scripts/.

Suggested-by: Petr Pavlu <ppavlu@suse.cz>
Suggested-by: Petr Tesařík <petesarik@suse.de>