git_sort: move configuration to a data file
The Head objects are not serializable. Use plain list and convert to
Head objects once the configuration is loaded.

With the repository list in a datafile it's possible to separate the
real configuration that evolves over time from test configuration that
needs to be the same as when the tests were written.

Fixes: 3544134fcbad ("git_sort: netdev remotes switched from master to main branch")