scripts/cve_tools/kss-dashboard: prefer scripts/bugzilla to the system version
We use our own patched version of redhat's python-bugzilla.
Unfortunately, the opensuse package is much closer to the vanilla
version and uses https://bugzilla.redhat.com as DEFAULT_BZ instead of
https://apibugzilla.suse.com.  Which means that if you install an open
suse version of python bugzilla on your system, this script stops
working, because it imports system version instead of the
kernel-source one.