blacklist.conf: CVE-2021-47289 bsc#1224984: not applicable, code does not exist
Patch adds a NULL pointer check to acpi_dev_put(), but this does not
exist in this codebase. It was added as an unrelated fix fe066621c7966f
("gpio: merrifield: Fix build err without CONFIG_ACPI") and then a lot
of refactoring in ACPI subsystem started to use it.

There isn't an equivalent code that should do the NULL pointer check,
probably added later in the refactoring as well as the bug.