Revert "Merge branch 'users/dwagner/SLE12-SP5/for-next' into SLE12-SP5"
This reverts commit aa4c39a920ecb484add5aa1733bbaa0fb81c7d46, reversing
changes made to 4527634da2625f9c0c83176368afe9fe8acb3ffc.

---

Following breaks kABI:

commit 72d636029eff5515a118fd98f44689c4421a836e
Author: Daniel Wagner <dwagner@suse.de>
Date:   Mon Sep 30 15:48:52 2024 +0200

   kabi: ignore all nvme kabi breakages

   Streamline sle12sp5 with the other code stream where we ignore
   all symbol changes inside the nvme subsystem.

   Delete:
    - patches.kabi/kabi-Fix-nvme-fabrics_q.patch
    - patches.kabi/kabi-Fix-nvmet-error-log-definitions.patch
    - patches.kabi/kabi-nvme-fix-fast_io_fail_tmo.patch

---

As designed the path match does not match symbols exported from vmlinux
(built-in), those have to be listed explicitly.

Listing the offending symbols should make this change work. It's
possible that more of the nvme support is modular on later kernels or
the kABI brekage is not as widespread compared to 4.12.

----