scripts/python/kss-dashboard: implement --stats-info and --sort-changes
An ability to show patch statistics and sort bugs based on the number
of inserted or delete lines.

The main usecase is to quickly identify likely candidates for the
patch monkey queue.  Even though the number of removed or added lines
is not the only metric, it's the only one that can be automated.