scripts/check-kernel-fix: refactor and clean up print_action
The current approach is to gather actions and warnings and print
it at the very end which brings some challenges - e.g. collecting
multiple warnings into a single variable is messy and that leads to code
duplication.

Refactor the code and print the action where it is examined (use
__print_action to do the ACTION NEEDED header) and remove the code
duplication introduced by blacklist.conf check which is now in its own
check_blacklisted_in_parent.

No functional changes intended.