scripts/check-kernel-fix: move cvss handling into check_branch_state
cvss filtering is currently done at the presentation layer
(find_and_print_toplevel_actions) which is suboptimal because we need to
do relatively expensive queries also for branches which will be always
filtered out as ineligible.

Move the filtering to check_branch_state as a first check. For the vast
majority of CVE fixes this will automatically rule out all cve and LTSS
branches which is generally a good thing.

Also teach cvss_affects_branch to handle empty cvss (usually the case
when $sha without any CVE is provided as an argument), we are continuing
to handle that as CVSS 0 to avoid considering LTSS branches.

The only corner case to note here. In the verbose mode we would lose
missing references check for branches which already do have fix even
though their scoring wouldn't apply to them. I am leaving that unplugged
here and follow up patch should deal with that for easier review and in
case we do not really care.