scripts/check-kernel-fix: drop quiet_mode
quiet_mode has only ever been used to print progress dots. As we usually
dump the c-k-f output to the bugzilla those are standing in the way and
the script is fast enough (mildly amusing right ;) the progress bar is
not strictly needed. We might add a parameter for people to get one if
they need.