scripts/check-kernel-fix: improve -v action reporting
ACTION NEEDED!
is reported anytime there is an actual thing to do. In verbose mode this
includes any action which is arguably dubious because there is simply
nothing to be done when the branch is not affected or it already
contains the fix or the fix has been blacklisted. Only use
__print_action for those actions that really require any activity.

This should make no effect on runs without -v parameter.