scripts/check-kernel-fix: cache the number of bugs for c-k-f purposes
The default cache expiry date is 2 hours for each email.

This should prevent c-k-f from spamming bugzilla we useless queries
and avoid rate-limiting issues when one runs c-k-f many time in rapid
succession.

One can always inspect the cache with...
$ jq <~/.cache/check-kernel-fix/bugzilla_cache.json

On c-k-f side it's not necessary to handle the error output, the new
script doesn't produce any except for a number (0 in case of an error)
and errors are differentiated purely by returning a different number.

As a side effect, this should make c-k-f slightly faster on average.