acpi/ghes: Process CXL Component Events (jsc#PED-10860).
This commit is a bit tricky. It looks very similar to 3e33eb3e6006 but
it's not. The initial patch to process CXL components had a bug which
resulted in its partial revert in the upstream kernel (this is reflected
by our 4bac2e52b1ff). However, certain definitions from the original
patch remained, so instead of trying to cobble up one final commit from
the 2 versions I'd rather follow the upstream history by backporting the
buggy patch, its partial revert as well as the final fixed version (this
commit).

The only unfortunate side effect is that both versions used the same
subject which results in identical file names for the backports. I've
simply added a -2 suffix to the final version.