io_uring: mockup implementation of futex operations (bsc#1230569).
We won't support io_uring futex operations in SLE15-SP7 because it would
require large changes to the futex code.  Nevertheless, operations in
io_uring are numbered sequentially and skipping some number is not
allowed.  Since we need to support newer operations, add a mock up
implementation of OP_FUTEX_{WAIT,WAKE,WAITV} that just returns
-EOPNOTSUPP, but which occupy the respective slots of these operations.

This patch must be dropped if we ever introduce the futex operation to a