scripts/check-kernel-fix: move warning evaluatoin to check_branch_state path
This will allow to run those in parallel which is not the case for
print_action. The only reason why this was originally in the
presentation layer was that those warnings are printed immediately
but we can store them into a file and print from there.

There shouldn't be any actual changes to the logic done by this move and
the only side effect is the runtime should be improved as some of those
warnings take time to evaluate (e.g. pending PRs need to grep all
for-next branches for the specific branchs).