MyBS: Use buildflags to set which package to build
Initially this was handled with package links and package meta.

Due to policy change package links are no longer allowed in packages
that are submitted to product releases.

The altearnative is multibuild. However, with multibuild package meta
cannot be used to set which spec file is built in which repository.

A workaround with setting macros in prjconf and using complex
conditional in the spec files aws added. However, the decision which
package is built in which repository belongs to the project
configuration, not the package.

New workaround is to use build flags in prjconf instead.