- Refresh   patches.suse/0002-iommu-vt-d-Fix-NULL-domain-on-device-release.patch. - Refresh   patches.suse/iommu-vt-d-Move-inline-helpers-to-header-files. - Delete patches.suse/kabi-padding-for-vmstat-items.patch.
The kernel-source commit 67017af3a15a wrongly deleted most of the
content in patch kabi-padding-for-vmstat-items.patch and only left an
unrelated commit. Since kABI is hard frozen now it can not be
restored.

But at least clean up the mess by folding the single hunk in this
patch to the commit where it belongs and remove the kABI patch.