scripts/common-functions: drop is_upstream_sha
is_upstream_sha is a misnomer because it only guarantees that the given
commit is in the referenced repository. It doesn't really check whether
it is reachable from a particular remote or branch. This is not a
problem for its only existing user because Fixes tags are referring to
upstream commits but the naming is misleading and more importantly we do
have a proper function for the purpose so use sha_in_upstream instead.