scripts/check-kernel-fix: wait for git-fixes background run properly
we are printing potential follow up fixes only if there is an action
required which is an intendeded behavior. We do want to wait for the run
to finish regardless of the final outcome though as we do not want the
git-fixes to outlive the script runtime. Theoretically we could just kill
git_fixes_pid but this could get more tricky if the process terminated
and the pid got recycled.