check-kernel-fix: check for config value inconsistencies in branch hierarchies
Currently the tool doesn't report a eligible branch because a related config
is disabled in one of the parents. We have cases where config is enabled in a
child, but disabled in parent. For example: CONFIG_AX25 is enabled in
SLE15-SP3-LTSS, but disabled in it's parent cve/linux-5.3-LTSS.
Report such anomalies and handle this corner case.

Also introduce two new branch states nope_unaffected_config and
unknown_config_disagree. Former indicates that branch has config
disabled, while latter indicates that config value is inconsistent with
one or more of it's children branches.