enabled CONFIG_X86_INTEL_TSX_MODE_AUTO
This is a response to bsc#1246695. As result of TAA vulnerability
(CVE-2019-11135) we have aimed to follow the upstream default for TSX
but due to a mistake we have ended up using CONFIG_X86_INTEL_TSX_MODE_ON
rather than CONFIG_X86_INTEL_TSX_MODE_OFF. This has been noticed later
on and fixed to align with upstream. Which has made some users unhappy
because they have lost a default TSX functionality even on HW that is
not susceptible to CVE-2019-11135.

We have discussed different ways to deal with that but the likely most
straightforward turned out to be to go with CONFIG_X86_INTEL_TSX_MODE_AUTO
which disables TSX only on CVE-2019-11135 affected HW. We are still
diverging from the upstream here but there are some positive indications
that no new TSX based side channels have been discovered since.