scripts/check-kernel-fix: list branches to blacklist for CVEs without known breaker
Quite a lot of CVEs are coming with unknown breakers and in some cases
it is simply too time consuming to identify them especially when it is
clear that this CVE would be blacklisted for some reason.

Our current
instructions suggest that people can use c-k-f -F to get a list of
branches that might need a fix but that takes some time to execute
and why to force people to run this if we already have that information.

Collect maybe_missing_commit into a file and print it with instructions
if a CVE is to be blacklisted.

Example

./scripts/check-kernel-fix CVE-2022-50359
Security fix for CVE-2022-50359 bsc#1250269 with CVSS 5.5
= 2b064d91440b ("media: cx88: Fix a null-ptr-deref bug in buffer_prepare()") merged v6.1-rc1~130^2~54
No Fixes tag. Requires manual review for affected branches.
Experts candidates:  tiwai@suse.com (58) subsystem/role="MEDIA DRIVERS"
Link: https://git.kernel.org/linus/2b064d91440b33fba5b452f2d1b31f13ae911d71
ACTION NEEDED!

Potential git-fixes for 2b064d91440b33fba5b452f2d1b31f13ae911d71
Nothing found
There is no Fixes tag for:
2b064d91440b ("media: cx88: Fix a null-ptr-deref bug in buffer_prepare()")
so we cannot automatically assess which kernel branches require the fix.
Please try to identify all the breakers and then
run ./scripts/check-kernel-fix -f BREAKER_SHA [-f BREAKER_SHA] CVE-2022-50359
to get the real list of branches to apply the fix into.

You can run ./scripts/check-kernel-fix -F CVE-2022-50359
to get an estimated list of branches to apply the fix into.

However, if the CVE turns out to be non-issue or an attack vector is not really feasible and
determining the breaker is not feasible then please blacklist in following branches:
SLE12-SP3-TD
SLE12-SP5