scripts/breaker_age: reimplement patch age discovery
note to self git log --reverse -1 doesn't way the way I would
anticipating/hoped for. It simply reduces number of commits visited
during the walk and print them in the reverse order. That means that I
do not see older commits :/

As we have regular update of references or context updates we cannot
limit the walk this way and then things are getting really time
consuming. I could find a reasonable way because everything just sucks.

This implementation abuses the structure of our patchfiles. The first
line is barely touched ever. We usually update references, changelog or
the patch body to refresh. git blame on the first line tends to be
a reliable source of the introducing commit. If we were to be really
precise --root should be used but this is slightly more costly in my
measurements.