kernel-binary: Do not change debuginfo config during build
Historically when debuginfo build was disabled in OBS kernel was
configured to not generate the debuginfo at all saving space during
build and making the build faster.

More and more kernel features depend on debuginfo, and disabling it
changes the kernel significantly disabling functionality that is
otherwise available and causing ABI breakage.

Recently genksyms was rewritten as gendwarfksyms to support more
features but requires debuginfo to operate. With that kernel builds
without deuginfo are not very useful anymore. Even if rpm eventually
trashes the debuginfo it needs to be always generated.