rpm/config.sh: Use suse-kabi-tools
Commit a2c6f8927eb8 ("packaging: Add support for suse-kabi-tools") updated
the kernel-binary spec template to enable the use of suse-kabi-tools for
kABI-related operations. This support is activated by setting
USE_SUSE_KABI_TOOLS=Yes in the rpm/config.sh file. Subsequently, commit
90af69e33ec2 ("rpm/config.sh: Use suse-kabi-tools (jsc#PED-12618)") enabled
the tools in SL-16.0 for kABI stability checking.

Enable now suse-kabi-tools in the master and its downstream branches as
well, so the old kABI scripts can eventually be removed. This branch
doesn't have a stable kABI, so only the 'ksymtypes consolidate' command is
used. The new tool is faster than the previous rpm/modversions script,
which speeds up the build process.