supported.conf: Update path for ufs drivers
As part of bsc#1253414 CVE-2025-40130, which updates
the ufs driver, it was discovered that the pathnames
in the supported module list had the old ufs driver
pathnames, which was drivers/scsi/ufs. But the
ufs drivers are now in drivers/ufs.

Also, the ti-j721e-ufs modules is now in the "host"
subdirectory.